major updates to quadrature grid
Overhaul of quadrature grid code and associated tests. In line with logic from mvQuad and tested for users providing their own quadrature rules. Sparse rule has been tested as well as pruning and made available in approxposterior. Not added officially to Laplace as this has minimal AGHQ options.

Laplace and Approx posterior were tweaked to accommodate this change. Additionally, I added the option of using an identity transformation (no transformation). It may be that some quadrules are not adaptive. This is incomplete in Laplace and needs more work (check weights etc).

In runNestedApprox, I updated `improveMarginals` to include the option to switch from AGHQ to a SPARSE grid or to prune.